+++ src/main/java/ru/itmo/product/repository/CoordinatesRepository.java
@@ -3,5 +3,5 @@ package ru.itmo.product.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.itmo.product.model.Coordinates;
 
-public interface CoordinatesRepository extends CrudRepository<Coordinates, Long> {
+public interface CoordinatesRepository extends CrudRepository<Coordinates, Integer> {
 }
