+++ src/main/java/ru/itmo/product/repository/OrganizationRepository.java
@@ -7,7 +7,7 @@ import ru.itmo.product.model.enums.OrganizationType;
 import java.util.List;
 import java.util.Optional;
 
-public interface OrganizationRepository extends CrudRepository<Organization, Long> {
+public interface OrganizationRepository extends CrudRepository<Organization, Integer> {
 
     Optional<Organization> findByNameAndFullNameAndAndEmployeesCountAndType(String name,
                                                                             String fullName,
