+++ src/main/java/ru/itmo/product/repository/ProductRepository.java
@@ -7,9 +7,9 @@ import ru.itmo.product.model.enums.UnitOfMeasure;
 
 import java.util.List;
 
-public interface ProductRepository extends CrudRepository<Product,Long> {
+public interface ProductRepository extends CrudRepository<Product,Integer> {
     void deleteAllByUnitOfMeasure(UnitOfMeasure unitOfMeasure);
-    List<Product> findAllByManufacturer_Id(Long id);
+    List<Product> findAllByManufacturer_Id(Integer id);
 
     @Query(value = "SELECT DISTINCT ON (manufacturecost) * FROM product", nativeQuery = true)
     List<Product> findDistinctByManufactureCost();
